scope imports, separate client out
maybe this will speed up vscode??
