+++ packages/api/router/system/index.ts
@@ -1,7 +1,7 @@
-import { adminProcedure, defineRouter } from '~api/lib/trpc'
+import { defineRouter } from '~api/lib/trpc'
 
 import { permissionSubRouter } from './permission'
-import { userRoleSubRouter } from './role'
+// import { userRoleSubRouter } from './role'
 
 export const systemRouter = defineRouter({
 	permissions: permissionSubRouter,
