+++ packages/api/trpc/ssr.ts
@@ -1,10 +1,11 @@
 import { createServerSideHelpers } from '@trpc/react-query/server'
 import { type GetServerSidePropsContext, type NextApiRequest, type NextApiResponse } from 'next'
 
-import { getServerSession, type Session } from '@weareinreach/auth'
-
-import { createContextInner, transformer } from '../lib'
-import { appRouter } from '../router'
+import { getServerSession } from '@weareinreach/auth/next-auth/get-session'
+import { type Session } from '@weareinreach/auth/next-auth/types'
+import { createContextInner } from '~api/lib/context'
+import { transformer } from '~api/lib/transformer'
+import { appRouter } from '~api/router'
 
 interface SSRContext {
 	req: GetServerSidePropsContext['req']
