+++ packages/auth/next-auth/types.ts
@@ -1,0 +1,1 @@
+export type * from 'next-auth'
