+++ packages/db/client/index.ts
@@ -1,0 +1,59 @@
+/* eslint-disable node/no-process-env */
+import { type Prisma, PrismaClient } from '@prisma/client'
+import { createPrismaQueryEventHandler } from 'prisma-query-log'
+import { Logger } from 'tslog'
+
+import { idMiddleware } from '~db/lib/idMiddleware'
+import { superjsonMiddleware } from '~db/lib/superjsonMiddleware'
+
+const log = new Logger({ name: 'prisma' })
+// eslint-disable-next-line turbo/no-undeclared-env-vars
+const verboseLogging = Boolean(process.env.NODE_ENV === 'development' && !!process.env.NEXT_VERBOSE)
+declare global {
+	// allow global `var` declarations
+	// eslint-disable-next-line no-var
+	var prisma: PrismaClient<typeof clientOptions> | undefined
+}
+
+const clientOptions = {
+	log: verboseLogging
+		? [
+				{ level: 'query', emit: 'event' },
+				{ level: 'error', emit: 'stdout' },
+				{ level: 'warn', emit: 'stdout' },
+		  ]
+		: [
+				{ level: 'error', emit: 'event' },
+				{ level: 'warn', emit: 'event' },
+		  ],
+} satisfies Prisma.PrismaClientOptions
+
+const prisma = global.prisma || new PrismaClient(clientOptions)
+
+prisma.$use(idMiddleware)
+prisma.$use(superjsonMiddleware)
+
+const queryLogger = createPrismaQueryEventHandler({
+	queryDuration: true,
+	format: true,
+	indent: `\t`,
+	// linesBetweenQueries: 2,
+	language: 'pl/sql',
+	logger: (data) => log.info(`\n${data}`),
+})
+
+if (!global.prisma) {
+	if (verboseLogging) {
+		prisma.$on('query', queryLogger)
+	} else {
+		prisma.$on('error', (event) => log.error(event))
+		prisma.$on('warn', (event) => log.warn(event))
+	}
+}
+
+if (process.env.NODE_ENV !== 'production') {
+	global.prisma = prisma
+}
+export { prisma }
+export type * from '@prisma/client'
+export { Prisma } from '@prisma/client'
